navajo/requirements.txt
requests
beautifulsoup4
playwright
pytesseract
pillow
groq
